+++ src/client/cypress/e2e/editor/lithologicalDescription.cy.js
@@ -44,11 +44,9 @@ describe("Tests for the lithological description column.", () => {
     // add lithological description
     cy.get('[data-cy="add-litho-desc-icon"]').click();
     cy.wait("@lithological_description");
-    cy.get('[data-cy="lithological-description-0"]').contains("0 m");
+    cy.get('[data-cy="description-0"]').contains("0 m");
 
-    cy.get(
-      '[data-cy="lithological-description-0"] [data-testid="ModeEditIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-0"] [data-testid="ModeEditIcon"] ').click();
     cy.get('[data-cy="description-textfield"]')
       .find("textarea")
       .first()
@@ -67,25 +65,19 @@ describe("Tests for the lithological description column.", () => {
       .click();
 
     // stop editing
-    cy.get(
-      '[data-cy="lithological-description-0"] [data-testid="ClearIcon"]',
-    ).click();
+    cy.get('[data-cy="description-0"] [data-testid="ClearIcon"]').click();
 
-    cy.get('[data-cy="lithological-description-0"]').contains("0 m");
-    cy.get('[data-cy="lithological-description-0"]').contains(
-      "A new description.",
-    );
-    cy.get('[data-cy="lithological-description-0"]').contains(
+    cy.get('[data-cy="description-0"]').contains("0 m");
+    cy.get('[data-cy="description-0"]').contains("A new description.");
+    cy.get('[data-cy="description-0"]').contains(
       "Quality of the description: very good",
     );
-    cy.get('[data-cy="lithological-description-0"]').contains("50 m");
+    cy.get('[data-cy="description-0"]').contains("50 m");
 
     // add lithological description that stretches two layers
     cy.get('[data-cy="add-litho-desc-icon"]').click();
     cy.wait("@lithological_description");
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="ModeEditIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="ModeEditIcon"] ').click();
 
     // fill to depth dropdown
     cy.get('[data-cy="to-depth-select"]')
@@ -99,19 +91,11 @@ describe("Tests for the lithological description column.", () => {
     cy.wait("@lithological_description");
 
     // stop editing
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="ClearIcon"]',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="ClearIcon"]').click();
 
-    cy.get('[data-cy="lithological-description-1"]').should(
-      "have.css",
-      "height",
-      "280px",
-    );
+    cy.get('[data-cy="description-1"]').should("have.css", "height", "280px");
     // delete last layer
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="DeleteIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="DeleteIcon"] ').click();
 
     cy.contains("Confirm").click();
     cy.wait("@lithological_description"); // delete request
@@ -121,54 +105,42 @@ describe("Tests for the lithological description column.", () => {
     // add two lithological descriptions
     cy.get('[data-cy="add-litho-desc-icon"]').click();
     cy.wait("@lithological_description");
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="ModeEditIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="ModeEditIcon"] ').click();
     cy.wait("@lithological_description");
 
     // stop editing
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="ClearIcon"]',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="ClearIcon"]').click();
     cy.wait("@lithological_description");
     cy.wait(2000);
 
     cy.get('[data-cy="add-litho-desc-icon"]').click();
     cy.wait("@lithological_description");
-    cy.get(
-      '[data-cy="lithological-description-2"] [data-testid="ModeEditIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-2"] [data-testid="ModeEditIcon"] ').click();
     cy.wait("@lithological_description");
 
     // stop editing
-    cy.get(
-      '[data-cy="lithological-description-2"] [data-testid="ClearIcon"]',
-    ).click();
+    cy.get('[data-cy="description-2"] [data-testid="ClearIcon"]').click();
     cy.wait("@lithological_description");
     cy.wait(2000);
 
     // delete the one in the middle
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="DeleteIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="DeleteIcon"] ').click();
 
     // assert error message
-    cy.get('[data-cy="lithological-description-1"]').contains(
+    cy.get('[data-cy="description-1"]').contains(
       "You are about to delete this layer, how do you want to proceed?",
     );
     cy.contains("Confirm").click();
     cy.wait("@lithological_description"); // delete request
     cy.wait("@lithological_description"); // updated get request
-    cy.get('[data-cy="lithological-description-1"]').contains(
+    cy.get('[data-cy="description-1"]').contains(
       "There is an undefined interval in the sequence",
     );
     // refill the middle
-    cy.get(
-      '[data-cy="lithological-description-1"] [data-testid="AddCircleIcon"] ',
-    ).click();
+    cy.get('[data-cy="description-1"] [data-testid="AddCircleIcon"] ').click();
 
     // assert no error message
-    cy.get('[data-cy="lithological-description-1"]').contains(
+    cy.get('[data-cy="description-1"]').contains(
       "Quality of the description: -",
     );