+++ src/client/src/api/fetchApiV2.js
@@ -34,18 +34,28 @@ export async function fetchApiV2(url, method, payload = null) {
   }
 }
 
+// layers
 export const fetchLayerById = async id =>
   await fetchApiV2(`layer/${id}`, "GET");
 
 export const fetchLayersByProfileId = async profileId =>
   await fetchApiV2(`layer?profileId=${profileId}`, "GET");
 
+export const updateLayer = async layer => {
+  // remove derived objects
+  delete layer.createdBy;
+  delete layer.updatedBy;
+  return await fetchApiV2("layer", "PUT", layer);
+};
+
+// codelists
 export const fetchAllCodeLists = async () =>
-  await fetchApiV2(`codelist`, "GET");
+  await fetchApiV2("codelist", "GET");
 
 export const updateCodeLists = async codelist =>
-  await fetchApiV2(`codelist`, "PUT", codelist);
+  await fetchApiV2("codelist", "PUT", codelist);
 
+// lithological descriptions
 export const fetchLithologicalDescriptionsByProfileId = async profileId => {
   return await fetchApiV2(
     `lithologicaldescription?stratigraphyId=${profileId}`,
@@ -55,7 +65,7 @@ export const fetchLithologicalDescriptionsByProfileId = async profileId => {
 
 export const addLithologicalDescription = async lithologicalDescription => {
   return await fetchApiV2(
-    `lithologicaldescription`,
+    "lithologicaldescription",
     "POST",
     lithologicalDescription,
   );
@@ -63,7 +73,7 @@ export const addLithologicalDescription = async lithologicalDescription => {
 
 export const updateLithologicalDescription = async lithologicalDescription => {
   return await fetchApiV2(
-    `lithologicaldescription`,
+    "lithologicaldescription",
     "PUT",
     lithologicalDescription,
   );
@@ -73,19 +83,34 @@ export const deleteLithologicalDescription = async id => {
   return await fetchApiV2(`lithologicaldescription?id=${id}`, "DELETE");
 };
 
-export const updateLayer = async layer => {
-  // remove derived objects
-  delete layer.createdBy;
-  delete layer.updatedBy;
-  return await fetchApiV2(`layer`, "PUT", layer);
+// facies descriptions
+export const fetchFaciesDescriptionsByProfileId = async profileId => {
+  return await fetchApiV2(
+    `faciesdescription?stratigraphyId=${profileId}`,
+    "GET",
+  );
 };
+
+export const addFaciesDescription = async faciesDescription => {
+  return await fetchApiV2("faciesdescription", "POST", faciesDescription);
+};
+
+export const updateFaciesDescription = async faciesDescription => {
+  return await fetchApiV2("faciesdescription", "PUT", faciesDescription);
+};
+
+export const deleteFaciesDescription = async id => {
+  return await fetchApiV2(`faciesdescription?id=${id}`, "DELETE");
+};
+
+export const fetchUsers = async () => await fetchApiV2("user", "GET");
 
 // Enable using react-query outputs across the application.
 
 // eslint-disable-next-line react-hooks/rules-of-hooks
 export const useDomains = () =>
   useQuery("domains", () => {
-    const domains = fetchApiV2(`codelist`, "GET");
+    const domains = fetchApiV2("codelist", "GET");
     return domains;
   });
 
@@ -94,11 +119,19 @@ export const useLayers = profileId =>
     return fetchLayersByProfileId(profileId);
   });
 
+export const lithologicalDescriptionQueryKey = "lithoDesc";
+
 export const useLithoDescription = selectedStratigraphyID =>
   useQuery({
-    queryKey: ["lithoDesc", selectedStratigraphyID],
+    queryKey: [lithologicalDescriptionQueryKey, selectedStratigraphyID],
     queryFn: () =>
       fetchLithologicalDescriptionsByProfileId(selectedStratigraphyID),
   });
 
-export const fetchUsers = async () => await fetchApiV2("user", "GET");
+export const faciesDescriptionQueryKey = "faciesDesc";
+
+export const useFaciesDescription = selectedStratigraphyID =>
+  useQuery({
+    queryKey: [faciesDescriptionQueryKey, selectedStratigraphyID],
+    queryFn: () => fetchFaciesDescriptionsByProfileId(selectedStratigraphyID),
+  });